GIT 提交信息：
功能: 重構聯絡表單並更新用戶介面

- 新增一個 `ContactFormModal` 組件，實現具有驗證和提交邏輯的聯繫表單
- 在 `AboutView` 中，增加一個訂閱新聞通訊的表單，具有表單驗證，並使用 `ContactFormModal`
- 在 `FAQView` 中，移除頁面內的聯繫表單，改為使用 `ContactFormModal`
- 更新 `AboutView` 中的按鈕，以打開 `ContactFormModal`
- 修改 `AboutView` 中的新聞通訊訂閱表單，包括條款和隱私政策連結的更變
- 移除 `FAQView` 中與之前頁面內聯繫表單相關的未使用導入和代碼
