+++ src/views/AboutView.vue
@@ -1,9 +1,66 @@
 <script setup>
+import { inject, ref } from 'vue';
+import axios from 'axios';
+import { useForm } from 'vee-validate';
+import * as yup from 'yup';
+import ContactFormModal from '@/components/ContactFormModal.vue';
 import aboutData from '@/data/about.json';
+
+const isLoading = ref(false);
 
 const {
   headers, features, services, team,
 } = aboutData;
+
+const isShowModal = ref(false);
+
+function closeModal() {
+  isShowModal.value = false;
+}
+function showModal() {
+  isShowModal.value = true;
+}
+
+const validationSchema = yup.object({
+  email: yup.string().email('請輸入正確的Email格式 example@example.com').required('請輸入您的E-Mail'),
+});
+
+const {
+  errors, defineField, values, resetForm,
+} = useForm({
+  initialValues: {
+    email: '',
+  },
+  validationSchema,
+});
+
+const [email, emailAttrs] = defineField('email');
+
+const swal = inject('$swal');
+
+const onSubmit = (async () => {
+  isLoading.value = true;
+  try {
+    const { data } = await axios.post(`${import.meta.env.VITE_APP_API_URL}/api/contact/subscribe`, {
+      email: values.email,
+    });
+    swal({
+      title: '訂閱成功',
+      text: data.message,
+      icon: 'success',
+    });
+    resetForm();
+    isLoading.value = false;
+  } catch (error) {
+    swal({
+      title: '訂閱失敗',
+      text: error.response.data.message,
+      icon: 'error',
+    });
+    isLoading.value = false;
+  }
+});
+
 </script>
 
 <template>
@@ -52,23 +109,26 @@ const {
         </p>
       </div>
       <div class="flex flex-col gap-4">
-        <input type="button" value="加入我們" class="px-6 py-3 text-white rounded-full bg-gray-1" />
+        <button type="button" class="px-6 py-3 text-white rounded-full bg-gray-1" @click="showModal">
+          加入我們
+        </button>
       </div>
     </section>
     <section class="flex flex-col items-center justify-center px-5 py-16 mb-6">
-      <form class="p-8 space-y-6 rounded-2xl bg-gray-4" @submit.prevent="">
+      <div class="p-8 space-y-6 rounded-2xl bg-gray-4">
         <div class="flex flex-col gap-3">
           <h2 class="mb-5 text-4xl font-bold">訂閱電子報</h2>
           <p>獲取最新資訊與熱門的投票議題</p>
         </div>
-        <div class="space-y-4">
+        <form class="space-y-4" @submit.prevent="onSubmit">
           <div class="flex flex-col gap-6">
-            <input type="email" placeholder="輸入你的Email" class="px-4 py-3 rounded-full" />
+            <input type="email" v-model="email" v-bind="emailAttrs" placeholder="輸入你的Email" class="px-4 py-3 rounded-full" required :disabled="isLoading" />
+            <p class="text-red-500" v-if="errors.email">{{ errors.email }}</p>
             <input type="submit" value="送出" class="px-6 py-3 text-white rounded-full bg-gray-1" />
           </div>
-          <p>點擊訂閱表示您同意我們 <RouterLink to="/privacy">使用條款與隱私政策</RouterLink>.</p>
-        </div>
-      </form>
+          <p>點擊訂閱表示您同意我們 <RouterLink to="/terms" class="underline transition duration-150 hover:text-primary" target="_blank">使用條款</RouterLink>與<RouterLink to="/privacy" class="underline transition duration-150 hover:text-primary" target="_blank">隱私政策</RouterLink>.</p>
+        </form>
+      </div>
     </section>
     <section class="flex flex-col items-center justify-center gap-8 py-16">
       <div class="flex flex-col gap-8 py-8 border-t md:gap-12" v-for="service in services" :key="service.title">
@@ -87,4 +147,5 @@ const {
       </div>
     </section>
   </main>
+  <ContactFormModal :isShowModal="isShowModal" :closeModal="closeModal" />
 </template>
